# wk5/api-documentation/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/albums/": {
      "get": {
        "tags": ["Albums"],
        "description": "Gets all albums from DB. Artist name included from foreign key",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": ["Albums"],
        "description": "Add a new album to the DB",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Ride the Lightning"
                },
                "artist": {
                  "type": "string",
                  "example": "Metallica"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Album created sucessfully"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/albums/{id}": {
      "get": {
        "tags": ["Albums"],
        "description": "Get a single album by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/artists/": {
      "get": {
        "tags": ["Artists"],
        "description": "Gets all artists from DB",
        "responses": {
          "200": {
            "description": "Receive a JSON of all artists"
          }
        }
      },
      "post": {
        "tags": ["Artists"],
        "description": "Create new DB entry for an artist",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "artist": {
                  "type": "string",
                  "example": "Amy Winehouse"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Added artist to DB successfully"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/artists/{id}": {
      "get": {
        "tags": ["Artists"],
        "description": "Get a single artist by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}